Attempts to repair leveldb after failed to open db
This change the will attempt to trigger eleveldb:repair operation,
if there was an error to open leveldb (db_open).
If repair was successful it will try to re-open db, otherwise original
error is returned.
If second open fails, there will be no another attempt to repair the db.

This should help with the cases where leveldb got corrupt after
powercut etc. Previous version would fail to start Riak, leaving it
to the Ops to find and repair corrupt leveldb. Ops would likely perform
the same procedure but doing it via Erlang shell.

New behaviour is turned on by default. Currently there is no way
to disable it (requires changes in the configuration module).